agui_server/pyproject.toml
[project]
name = "agui-adk-a2a-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ag-ui-protocol>=0.1.9",
    "aiofiles>=25.1.0",
    "google-adk[a2a]>=1.15.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
    "aiosqlite>=0.21.0",
    "sqlalchemy[asyncio]>=2.0.43",
]
